requirements.txt
discord.py>=2.3
watchdog>=3.0
requests>=2.31
python-dotenv>=1.0
aiohttp>=3.9
aiosqlite>=0.19
